Update Debugger.php
By default the folder of controllers is in resources, but I think that it's more suitable place in the folder app / controller. Therefore, I use the existing sober / controller / path filter with the value get_theme_file_path (). '/ App / controller', and everything seems to work, but it does not suit the fact that the output in the debugger turns into something like /Users/user/Sites/site/web/app/themes/my-theme/app/controller/file .php. A modified line of code solves this problem.